service2/s2_requirements.txt
# Flask
Flask==1.1.2
python-dotenv==0.14.0

# APIs
requests==2.24.0

# Python Code Testing
pytest==6.0.1
pytest-mock==3.2.0

# Violation Checking in Jenkins
pylint==2.5.3
pylint-fail-under==0.3.0
pycodestyle==2.6.0
